Cleanups
Using size_t instead of int throughout the code
Using C++11 initializer syntax
Random subset and random mask algorithms that handle more than 4G elements

+++ JrBoostLib/AbstractPredictor.h
@@ -6,7 +6,7 @@ protected:
 
 public:
     virtual ~AbstractPredictor() = default;
-    virtual int variableCount() const = 0;
+    virtual size_t variableCount() const = 0;
     virtual ArrayXf predict(const Eigen::ArrayXXf& inData) const = 0;
     
 // deleted:
